Move AmazonSageMakerTrainingLogsService to training
