+++ app/tests/components_tests/test_models.py
@@ -12,7 +12,7 @@ from lambda_tasks.timeouts import SoftTimeLimitExceeded
 from grandchallenge.algorithms.models import AlgorithmImage, Job
 from grandchallenge.cases.models import Image
 from grandchallenge.cases.widgets import DICOMUploadWithName
-from grandchallenge.components.backends.amazon_sagemaker_base import (
+from grandchallenge.components.backends.amazon_sagemaker_training import (
     AmazonSageMakerTrainingLogsService,
 )
 from grandchallenge.components.models import (
@@ -957,7 +957,7 @@ def test_runtime_metrics_chart(mocker):
     mock_service.get_task_logs.return_value = []
 
     mocker.patch(
-        "grandchallenge.components.backends.amazon_sagemaker_base.AmazonSageMakerTrainingLogsService",
+        "grandchallenge.components.backends.amazon_sagemaker_training.AmazonSageMakerTrainingLogsService",
         return_value=mock_service,
     )
